+++ settings.advanced.inc.php
@@ -79,6 +79,9 @@ $REX['ADDON']['seo42']['settings']['static_sitemap_priority'] = true;
 // you can force download of certain filetypes. put file in files dir, add filetype to array e.g. 'pdf' and link to file like this: /download/foo.pdf or use seo42::getDownloadFile($file)
 $REX['ADDON']['seo42']['settings']['force_download_for_filetypes'] = array();
 
+// cache control header for image manager files is normally set through .htaccess, but on some servers (1und1) it won't be set correctly so we do it manually
+$REX['ADDON']['seo42']['settings']['fix_image_manager_cache_control_header'] = false;
+
 // if false seo database fields won't be dropped if seo42 will be uninstalled. perhaps someday interesting when updateing seo42...
 $REX['ADDON']['seo42']['settings']['drop_dbfields_on_uninstall'] = true;